PAY-5639 Fixed Sonar issues
